fix(UT): remove legacy column from identifier. ref: #30243 (#30244)
During the migration of customers from another database engine to
Postgres, a column was inadvertently left behind. This oversight has
resulted in 404 errors because in Postgres, the column is interpreted as
a function, not a column.

Added a note in the MainSuite2a since it lasts twice than other suites.