Issue 30291 osgi actionlet reinit failure (#30292)
This pull request forces OSGI to init BEFORE an actionlet is registered.
It changes to the `WorkflowAPIImpl` class to improve OSGI initialization
and error handling. The most important changes include adding a new
import for `OSGIUtil`, modifying the `registerBundleService` method to
ensure OSGI is initialized, and adding error handling for null
`BundleContext`.

### Improvements to OSGI Initialization and Error Handling:

*
[`dotCMS/src/main/java/com/dotmarketing/portlets/workflows/business/WorkflowAPIImpl.java`](diffhunk://#diff-5a5ef615df4ac7495073609cffd31eabd85513c4c81dd48cc82b5c6fd0523934R143):
Added import for `org.apache.felix.framework.OSGIUtil`.
*
[`dotCMS/src/main/java/com/dotmarketing/portlets/workflows/business/WorkflowAPIImpl.java`](diffhunk://#diff-5a5ef615df4ac7495073609cffd31eabd85513c4c81dd48cc82b5c6fd0523934L370-R387):
Modified `registerBundleService` method to initialize OSGI framework and
added error handling for cases where OSGI is not initialized or
`BundleContext` is null.