fix(UVE): UrlContentMap Redirection to Correct Page After Save or Publish #29896 (#30334)
### Proposed Changes
* Refactored the logic to handle URL redirection after saving a
UrlContentMap.
* Added methods to extract and determine the target URL for redirection,
ensuring the correct URL is used when publishing content.

### Checklist
- [x] Tests
- [x] Translations
- [x] Security Implications Contemplated (add notes if applicable)

### Additional Info
This fix addresses the issue where, after saving or publishing a
UrlContentMap, the system incorrectly redirects to the content type’s
detail page instead of the expected URL. The changes ensure that the
correct URL is detected and used, preventing unwanted redirection after
the save or publish action.

### Screenshots
#### Before:


https://github.com/user-attachments/assets/1b4f74fa-f0b7-46f4-b389-a1bc63f52f7b

#### After:


https://github.com/user-attachments/assets/84ea8aa1-a434-4c17-a4b3-b4f391a58d14