fix(tests): Several Tests where broken due new dotcms.com site ref:#30420 (#30422)
This pull request includes a small change to the `CircuitBreakerUrlTest`
test case. The change updates the assertion to check for a different
string in the response content.

*
[`dotcms-integration/src/test/java/com/dotcms/http/CircuitBreakerUrlTest.java`](diffhunk://#diff-a86d291af87dcbb5dfa27d7faa1479a6ca1237bc77d89e38dba28f1cb29d4c26L379-R379):
Updated the assertion in `public void testGet()` to check for the string
"dotCMS empowers brands to simplify content creation" instead of
"/application/themes/dotcms/js/bootstrap.min.js".