Issue 29238 Fix upload new files error in Tasks portlet (#30416)
The
[error](https://github.com/dotCMS/core/issues/29238#issuecomment-2392303924)
mentioned already existed before the fix of the original issue.
 
When uploading a file, the method made some checks to verify where was
the request coming from. This check wasn't considering the case when it
comes from the Task Detail tab (Attach File button), and it was trying
to set a language attribute (_CONTENT_SELECTED_LANGUAGE_) that was only
being set when we opened a contentlet before. So if we tried to upload a
file from the Task Detail tab before that, it was throwing a null-caused
exception. This was fixed by adding an extra check, if that attribute
was null then set another one that is present in all the cases where
this method is called. This way we won't get a case where it is null.

Co-authored-by: erickgonzalez <erick.gonzalez@dotcms.com>