fix(Content): Title override on Assets json transformation #30127 (#30408)
### Proposed Changes
* Updated `DotAssetViewStrategy` to avoid overwriting the `title` field
in the transformation map.
* Replaced `map.put(TITLE_FIELD, fileName)` with
`map.computeIfAbsent(TITLE_FIELD, k -> fileName)` to ensure the `title`
is only set if it is not already present.

### Checklist
- [x] Tests
- [x] Translations
- [x] Security Implications Contemplated (add notes if applicable)

### Additional Info
This fix prevents the `title` field from being overwritten when a
`DotAsset` is transformed. The issue occurred because the `title` field
was being replaced with the asset’s filename, even if it already had a
value.

### Video

https://github.com/user-attachments/assets/2fa860c5-e1c7-4d32-912d-eec40fe0fbc5