chore: feedback message. ref: #30156 (#30449)
This pull request includes updates to the `Language.properties` file to
improve the clarity and consistency of messages related to the
end-of-life (EOL) status of dotCMS versions.

Updates to EOL messages:

*
[`dotCMS/src/main/webapp/WEB-INF/messages/Language.properties`](diffhunk://#diff-21db9723fa4c52389731b7575f1e19bd364aa992b7e549e22c08a33456994e94L5811-R5812):
Updated the `lts.expired.message` to clarify that the version is no
longer supported and to use the term "customer success manager" instead
of "CSM".
*
[`dotCMS/src/main/webapp/WEB-INF/messages/Language.properties`](diffhunk://#diff-21db9723fa4c52389731b7575f1e19bd364aa992b7e549e22c08a33456994e94L5811-R5812):
Updated the `lts.expires.soon.message` to indicate that the version will
no longer be supported and to use the term "customer success manager"
instead of "CSM".