Issue 30023 Content added as the default fileAsset content type regardless of folder settings (#30451)
The main issue was happening because when uploading a File, it was
setting the default type (File) despite the one specified in the folder.
In other words, it was ignoring the default type that was set in the
folder before and setting the one of the contentlet that was being
edited.

As analyzed, the only place where the File type is specified when
uploading is from the site browser, in all the other cases you can't
choose what type of File you want to upload so it should take the
default one. That said, a fix was added to consider those cases, so now
when uploading a File, it is going to have the content type specified in
the parent folder.