fix: Avoid dupe params on query parameters (#30414)
This pull request makes several important changes to the
`VanityUrlRequestWrapper` class to improve the handling of query strings
and adds a corresponding test to ensure the new functionality works
correctly. The key changes include adding a utility method to convert
URL parameters to a map, modifying how query strings are merged, and
introducing a new test case.

### Improvements to Query String Handling:

* Modified the logic for merging query strings to ensure that vanity
parameters are only added if they are not already present in the request
parameters.
* Added a new private method `convertURLParamsStringToMap` to convert a
URL parameters string into a map of key-value pairs.

### Testing Enhancements:

* Introduced a new test case
`test_that_vanityUrlParams_requestParams_Are_Same_Should_Not_Be_Duped`
to verify that query string parameters are not duplicated when they are
the same in both the vanity URL and the request.