chore(Telemetry) #30079 : Move telemetry plugin into core - Part 1 (#30520)
### Proposed Changes
* Migrates all the code from the `Telemetry` plug-in into our `core`
project.
* This is the first part of the process, so it just takes the code "as
is" from the plugin, and places it inside the `core` project. It only
includes a few adjustments on some Javadoc, code formatting, and the
addition of the `toString()` method in a few classes.
* These changes can safely co-exist with the current `Telemetry` plugin.
* In order to have a more effective Code Review process, the goal is to
have two pull requests:
  * The first one with the existing code from the plugin -- this one.
* The second one with the code changes in `core` that effectively add
the REST Endpoint, Web Interceptor, and Quartz Job which are currently
being added via the `Activator` class.