fix(Unable to run IT from the IDE) Refs: #30564 (#30565)
Normally, the rule for running an integration test is to annotate it
with the JUnit4WeldRunner. Additionally, annotate the test as
@ApplicationScoped. However, there was an issue with a bean that was
instantiated deep within the core of the CMS, causing a failure that
impacted integration tests.
Particularly on `UniqueFieldValidationStrategyResolver`which has an
injection problem
I resolved the problem by moving the injection to the constructor

Injecting private fields as follows is problematic :

```
@Inject
private  DBUniqueFieldValidationStrategy dbUniqueFieldValidationStrategy;

```
This problem did not surface in other scenarios, but it became visible
when running directly from the integration test. Additionally, there was
a public field with an @Rule in JUnit 4 that was also causing issues,
which has been corrected as well.