chore(deployment): Update glowroot default temp and logging configuration (#30024) (#30560)
This pull request introduces several changes to the `dotCMS` project,
focusing on configuration updates, new development commands, and
improvements to the build process. The most important changes include
the addition of a new configuration file for Glowroot, updates to
development commands in the `justfile`, and enhancements to the
`parent/pom.xml` file to support dynamic port configurations.

### Configuration Updates:
* Added a new Glowroot configuration file `glowroot.logback.xml` to the
`dotCMS/pom.xml` file.
* Introduced the `glowroot.logback.xml` configuration file with detailed
logging settings.

### Development Commands:
* Updated `justfile` to update `dev-run` command for running the
application with Glowroot enabled and renamed the existing `dev-run`
that added debug without suspend to `dev-run-debug`.
* Removed the `dev-start` command from the `justfile` as was confusing
between dev-start and dev-run

### Build Process Enhancements:
* Modified `parent/pom.xml` to dynamically output the Glowroot port if
the profile is enabled.

### Docker Environment Updates:

* Made glowroot use the `CATALINA_TMPDIR` environment variable by
default for consistent behaviour. We may consider later changing to use
a DOTCMS_TMPDIR and set CATALINA_TMPDIR and others to this at a later
time.


when using maven and the just commands specifying
-Ddocker.glowroot.enabled=true will enable glowroot on a dynamic local
port when starting up to prevent port conflicts. Maven will report the
full url to glowroot ui. This has been added by default in the just file
to the dev-run command but can be added to any maven command that starts
up the dotcms image. The local docker port mapping can be fixed using
-Dglowroot.port=4000 if required.

Whe using a development docker-compose use the following.
GLOWROOT_WEB_UI_ENABLED is required to set the listen address to 0.0.0.0
to allow access.

      GLOWROOT_ENABLED: 'true'
      GLOWROOT_WEB_UI_ENABLED: 'true'
      

In the K8s env we need to set the remote collector address and
GLOWROOT_AGENT_ID e.g.

```
        - name: GLOWROOT_ENABLED
          value: 'true'
        - name: GLOWROOT_AGENT_ID
          value: dotcms-qa::master
        - name: GLOWROOT_COLLECTOR_ADDRESS
          value: http://glowrootcentral.dotcmscloud.com:8181
```

Related to #30024 (Need to ship with the updated glowroot.jar).


### Screenshots
Original             |  Updated
:-------------------------:|:-------------------------:
** original screenshot **  |  ** updated screenshot **