fix(pp) fixing edge push cases where content and folders from non-pushed hosts gets dragged along by dependency, and containers in a drafted state fail to publish (#30380)
ref: #30370

This PR updates the `FolderHandler` and the `ContentletHandler` to
ignore folders and contents that live on a host that does not exist in
the target system.

It also insures that the working inode of any pushed container is the
one that is just saved by the `ContainerHandler` when it is creating a
container for the first time, which fixes a case where there are
different live/working versions of a container on the sender and the
sender only sends the live version.

Finally, it fixes the crappy Bundle Upload screen so now the dialog does
not stay open the whole time while the bundle is being uploaded AND
ingested.