chore(uve): Add sort navigation to traditional and nextjs example (#30624)
Showcase:


https://github.com/user-attachments/assets/49551094-9a30-4482-8765-7bf2a37693e2


This pull request introduces significant changes to the
`EditEmaEditorComponent` and related utilities to improve the handling
of menu reordering. The most important changes include the addition of a
new utility function for creating reorder menu URLs, updates to the
`EditEmaEditorComponent` to use this new function, and modifications to
the SDK to support the new reorder functionality.

### Enhancements to menu reordering functionality:

*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/utils/index.ts`](diffhunk://#diff-2fd4d225bb33327848da73766ec5efbe037e3aed1a1e51a50c19f6a2be8cdb80R516-R558):
Added the `createReorderMenuURL` function to construct URLs for the
reorder menu page with specified parameters.
*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/edit-ema-editor.component.ts`](diffhunk://#diff-24dc496db1eb6feb3e10a031baa4b4b63c20f1cd02ade574065f6b967f11c365L998-R1017):
Updated the `CLIENT_ACTIONS.REORDER_MENU` action to use the new
`createReorderMenuURL` function for generating the reorder URL.

### SDK updates:

*
[`core-web/libs/sdk/client/src/lib/editor/sdk-editor.ts`](diffhunk://#diff-da7281c94a8714f497eefea7843cb58b1b25aec7513e621a0214694697ac4034R46-R64):
Added the `reorderMenu` function to the SDK to handle menu reordering by
sending the appropriate message to the editor.
*
[`core-web/libs/sdk/client/src/index.ts`](diffhunk://#diff-e1656e0e6ea2bbbf1cd43fb6b12f3e4c419d28dd7d11401f31facf5b3f4eb75aR24):
Exported the new `reorderMenu` function from the SDK.

### Component and test updates:

*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/edit-ema-editor.component.ts`](diffhunk://#diff-24dc496db1eb6feb3e10a031baa4b4b63c20f1cd02ade574065f6b967f11c365R805-R813):
Added a temporary solution to reload the content by reloading the
window.
*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/utils/utils.spec.ts`](diffhunk://#diff-6b95047786d13d69ebdcb642c6d382bc35784a2ada53f95d1908119ab08dfb91R713-R740):
Added tests for the `createReorderMenuURL` function to ensure it
correctly adds missing query parameters.

### Velocity template updates:

*
[`dotCMS/src/main/webapp/WEB-INF/velocity/VM_global_library.vm`](diffhunk://#diff-9623596572dc2aec2ab2e750493bacb7cd3eaf5790edc4541a1cc65e84230e4fL175-R186):
Replaced the inline JavaScript for opening the reorder dialog with calls
to the new `reorderMenu` function.
[[1]](diffhunk://#diff-9623596572dc2aec2ab2e750493bacb7cd3eaf5790edc4541a1cc65e84230e4fL175-R186)
[[2]](diffhunk://#diff-9623596572dc2aec2ab2e750493bacb7cd3eaf5790edc4541a1cc65e84230e4fL226-L246)

---------

Co-authored-by: Kevin Davila <kfariid@gmail.com>