fix(examples): Hide edit button outside editor (#30657)
This pull request includes changes to the
`examples/nextjs/src/components/shared/contentlets.js` file to improve
the functionality and performance of the `Contentlets` component. The
most important changes include replacing `React` with `useMemo`, adding
a new import for `isInsideEditor`, and conditionally rendering the edit
button based on whether the contentlet is inside the editor.

Improvements to functionality:

*
[`examples/nextjs/src/components/shared/contentlets.js`](diffhunk://#diff-b6f3603978bcab24fcdd25b0a658c2ae69d34c68a14f6f5be88a0c1fb0e8a0edL2-R4):
Replaced the `React` import with `useMemo` to memoize the
`isInsideEditor` function.
*
[`examples/nextjs/src/components/shared/contentlets.js`](diffhunk://#diff-b6f3603978bcab24fcdd25b0a658c2ae69d34c68a14f6f5be88a0c1fb0e8a0edL2-R4):
Added a new import for `isInsideEditor` from `@dotcms/client` to check
if the contentlet is inside the editor.
*
[`examples/nextjs/src/components/shared/contentlets.js`](diffhunk://#diff-b6f3603978bcab24fcdd25b0a658c2ae69d34c68a14f6f5be88a0c1fb0e8a0edR13-R19):
Used `useMemo` to create a memoized value `insideEditor` for determining
if the contentlet is inside the editor.
*
[`examples/nextjs/src/components/shared/contentlets.js`](diffhunk://#diff-b6f3603978bcab24fcdd25b0a658c2ae69d34c68a14f6f5be88a0c1fb0e8a0edR37):
Conditionally rendered the edit button based on the `insideEditor` value
to improve user experience.