fix(block editor): Support UrlContentMap on Links (#30674)
This pull request includes a change to the `BubbleLinkFormComponent`
class in the
`core-web/libs/block-editor/src/lib/extensions/bubble-link-form/bubble-link-form.component.ts`
file. The change focuses on cleaning up the URLs for contentlets.

Improvements to URL handling:

*
[`core-web/libs/block-editor/src/lib/extensions/bubble-link-form/bubble-link-form.component.ts`](diffhunk://#diff-dd9a756d839455372ca42ef647e07ce405f4e476065c152bdb06c5bc3f6f8259R248-R263):
Added logic to replace the URL in `contentlet` with the actual URL
stored in `urlMap` if `URL_MAP_FOR_CONTENT` is present. This ensures
that the URLs for `urlContentMaps` are correctly formatted.