feat(UVE): Bring back inline editing for block editor fields in Headless (#30638)
This pull request introduces significant changes to the inline editing
functionality within the `core-web` library, primarily focusing on the
block editor and WYSIWYG editor integration. The updates include
modifications to the event handling, data structures, and messaging
between components. Below are the most important changes grouped by
theme:

### Inline Editing Event Handling:

* Introduced a new `#handleInlineEditingEvent` method in
`EditEmaEditorComponent` to handle different types of inline editing
events, such as `BLOCK_EDITOR` and `WYSIWYG`. This method checks for
enterprise licenses and handles the event accordingly.
* Updated the `CLIENT_ACTIONS` enum to remove
`INIT_BLOCK_EDITOR_INLINE_EDITING` and added `INIT_INLINE_EDITING` to
streamline the event handling process.

### Data Structure Updates:

* Modified the `InlineEventData` interface to replace
`blockEditorContent` with a structured `content` object in
`dot-block-editor-sidebar.component.ts`. This change ensures that the
content data is more organized and easier to manage.
* Updated the `BlockEditorData` interface to include a `JSONContent`
type for the `content` field, enhancing the data consistency and type
safety.

### Messaging and Communication:

* Refactored the `initInlineEditing` function to send messages to the
editor with the appropriate payload for both `BLOCK_EDITOR` and
`WYSIWYG` types. This change standardizes the communication between the
client and the editor.
* Added tests for the `initInlineEditing` function to ensure it sends
the correct messages based on the provided type and data.

### Component and Service Adjustments:

* Removed the `#getJsonContent` method from
`DotBlockEditorSidebarComponent` as it is no longer needed with the
updated `content` handling.
* Updated the `InlineEditService` to use the new
`CLIENT_ACTIONS.INIT_INLINE_EDITING` action, ensuring consistent event
handling across the application.

### Test Updates:

* Modified existing tests and added new ones to cover the changes in
event handling, data structures, and messaging. These updates ensure
that the new functionality is thoroughly tested and reliable.
[[1]](diffhunk://#diff-91349cd92f693c5f3ef517eb5ab5e553167c142d330bce29621b892583b2c177L14-R41)
[[2]](diffhunk://#diff-bae6613453ca5e03886194e908b1dbed1ffe067808be88cd178dbc8defa024ccR128-R164)

### Headless


https://github.com/user-attachments/assets/4b8dc731-f805-46f2-8910-05485112b174

### Traditional


https://github.com/user-attachments/assets/66a8967a-438c-4b45-9c78-f052326ac872


### WYSIWYG


https://github.com/user-attachments/assets/c4d2087c-932b-43ed-ba5f-1b6c4bed0fc5