fix(db)  allow db dump functionality to work even if the password val… (#30744)
This pull request includes updates to the `DataSourceAttributes` class
in the `dotCMS` project to improve the handling of username and password
encoding. The most important changes are:

Improvements to encoding:

*
[`dotCMS/src/main/java/com/dotcms/util/DataSourceAttributes.java`](diffhunk://#diff-09f7a1123aaff454ae6e8a9fef7c27f993745035d90252c8de24ba25f6b1c912R4-R7):
Imported `Try` from `io.vavr.control`, `URLEncoder`, and
`StandardCharsets` to handle encoding.
*
[`dotCMS/src/main/java/com/dotcms/util/DataSourceAttributes.java`](diffhunk://#diff-09f7a1123aaff454ae6e8a9fef7c27f993745035d90252c8de24ba25f6b1c912L16-R21):
Updated the constructor to encode the `username` and `password` using
`URLEncoder` and `StandardCharsets.UTF_8`, and handle potential encoding
exceptions using `Try`.