Issue 29018 Relationship Field showFields Field Variable value showing identifier instead of id field (#30749)
It seems that when retrieving the data of the related contentlet, an id
field is being set to avoid the UI to get broken `Duplicates value for
identifier key in map so that UI does not get broken`.

So to keep this behaviour a workaround was applied to solve the case in
which a field id was added to the key **showFields** in relationships
field field variable.

The workaround checks whether the field added to showFields is "id", if
true, then it goes to find the value of that field to **siblings**
field, where the real value of the field is going to be set. If not
found, then set the contentlet identifier.

<img width="1239" alt="Captura de pantalla 2024-11-22 a la(s) 4 11 52 p
 m"
src="https://github.com/user-attachments/assets/8832d454-9f04-42c3-9409-fc6e3fe218e8">